Add console support for runX
Add support for console redirect from Xen to containerd in the runX
scripts.  This does introduce a new outside dependency on daemonize but
that is needed to get socat to work (other wise socat kills itself).  A
basic timeline of what happens is below:
  ...
   |
   v
create <- we need to create a tty to give to containerd but Xen hasn't
   |      started yet so we create a fake with socat and pass that off
   |      with serial_fd_handler.  Also, we use recvtty to trick runc
   |      into thinking it's connected to containerd.
   v
start  <- after Xen creates the VM, we use socat again to connect the
   |      pts xenstore gives us to a unix socket that was created with
   |      the create usage of socat.  This connects Xen's console to
   |      containerd and is available to docker or other software.
   V
  ...

Signed-off-by: Brian Woods <brian.woods@xilinx.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>