Use a variable to define the busybox binary to use
The busybox binary to use with runX needs to be statically built for the
target architecture. However, it is also run at built time by
make-initrd to create links inside the initrd. If runX is been
cross-compiled, a QEMU user binary needs to be installed on the build
host.

No functional changes.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>