Fix delete script to work with containerd
It seems that containerd first has to kill the container process (xl
destroy in runX's case) before it can delete (gets rid of the container
FS).   So change running the delete script to when kill is called.  Also
use a PID for recvtty so that we can kill it as well.

Signed-off-by: Brian Woods <brian.woods@xilinx.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>