Add busybox (cross)build for arm64
So far the busybox binary to use for the domU initrd has to be
externally provided. While we still support that, also add the ability
to fetch busybox and cross-compile it with the right config on arm64.

If the environmental variable "busybox" is set, then use the provided
busybox, if not, then fetch and build busybox ourselves.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Acked-by: Bruce Ashfield <bruce.ashfield@xilinx.com>
Tested-by: Bruce Ashfield <bruce.ashfield@xilinx.com>