Don't call busybox --install
Busybox --install is convenient but has the problem that it only works
for native builds. If busybox is of a different arch compared to the
target, it cannot be run at build time with assistance.

Solve the problem by installing busybox by hand, creating links to the
applets from the make-initrd script.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Acked-by: Bruce Ashfield <bruce.ashfield@xilinx.com>
Tested-by: Bruce Ashfield <bruce.ashfield@xilinx.com>