automation: switch to an x86 runner
The CI-loop is switching to an x86 runner. For it to work we need to:

- cross-build RunX in the build stage
- run qemu-system-aarch64 from an x86 environment in the test stage
- export the aarch64 test-artifacts on an x86 host

Switch to debian:unstable for the build job and the test job because
Debian comes with an aarch64 cross-compiler and with a
qemu-system-aarch64 that works out of the box.

For the test-artifacts targets, add an x86 busybox inside the
test-artifacts containers. It can be used to run small scripts inside
those containers on x86, enough to export the content as artifacts.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>